[Race condition] Fix race condition part 1
